api_project-1/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-authtoken==3.14.0
python-dotenv==1.0.0